feat : step0 피드백 반영 & 코드 리팩토링
