+++ src/api/apiClient.ts
@@ -11,7 +11,7 @@ export interface ApiSuccess<T> {
 export const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
 
 const apiClient: AxiosInstance = axios.create({
-  baseURL: `${API_BASE_URL}`,
+  baseURL: API_BASE_URL,
   headers: { 'Content-Type': 'application/json' },
 });
