--- src/page/Home/components/GiftRankingSection.tsx
@@ -1,236 +1,0 @@
-import styled from '@emotion/styled';
-import { filters, generations } from '@/data/categoryDatas';
-import useSearchParamState from '../hooks/useSearchParamState';
-import useToggleCollapse from '../hooks/useToggleCollapse';
-import { generatePath, useNavigate } from 'react-router-dom';
-import { ROUTE_PATH } from '@/routes/routePath';
-import Loading from '@/components/Loading';
-import useRanking from '../hooks/useRanking';
-import toLocaleString from '@/utils/toLocaleString';
-
-interface ButtonProps {
-  isActive: boolean;
-}
-
-const GiftRankingSection = () => {
-  const {
-    handleGenerationGroupClick,
-    handleFilterGroupClick,
-    activeGenerationButton,
-    activeFilterButton,
-  } = useSearchParamState();
-
-  const { rankingDatas, loading } = useRanking({ activeGenerationButton, activeFilterButton });
-
-  const { isCollapsed, visibleItemsCount, toggleCollapse } = useToggleCollapse(rankingDatas.length);
-
-  const navigate = useNavigate();
-  const handleItemClick = (id: number) => {
-    navigate(generatePath(ROUTE_PATH.ORDER, { id: String(id) }));
-  };
-
-  const renderContent = () => {
-    if (loading) return <Loading />;
-    if (rankingDatas.length === 0) {
-      return <NoDataMessage>상품이 없습니다.</NoDataMessage>;
-    }
-    return (
-      <>
-        <RankContainer>
-          {rankingDatas.slice(0, visibleItemsCount).map((rank, index) => (
-            <RankItem key={rank.id} onClick={() => handleItemClick(rank.id)}>
-              <RankNumber>{index + 1}</RankNumber>
-              <ItemContainer>
-                <Image src={rank.imageURL} alt={rank.name} />
-                <ItemBrandName>{rank.brandInfo.name}</ItemBrandName>
-                <ItemName>{rank.name}</ItemName>
-                <ItemPrice>{toLocaleString(rank.price.basicPrice)} 원</ItemPrice>
-              </ItemContainer>
-            </RankItem>
-          ))}
-        </RankContainer>
-        <ToggleButton onClick={toggleCollapse}>{isCollapsed ? '펼치기' : '접기'}</ToggleButton>
-      </>
-    );
-  };
-
-  return (
-    <Section>
-      <Title>실시간 급상승 선물랭킹</Title>
-
-      <CatContainer>
-        <GenerationGroup>
-          {generations.map(({ id, emoji, label }) => (
-            <Button
-              key={id}
-              isActive={activeGenerationButton === id}
-              onClick={() => handleGenerationGroupClick(id)}
-            >
-              <div>{emoji}</div>
-              <p>{label}</p>
-            </Button>
-          ))}
-        </GenerationGroup>
-
-        <FilterGroup>
-          {filters.map(({ id, label }) => (
-            <Button
-              key={id}
-              isActive={activeFilterButton === id}
-              onClick={() => handleFilterGroupClick(id)}
-            >
-              <p>{label}</p>
-            </Button>
-          ))}
-        </FilterGroup>
-      </CatContainer>
-      {renderContent()}
-    </Section>
-  );
-};
-
-export default GiftRankingSection;
-
-const NoDataMessage = styled.div`
-  grid-column: 1 / -1;
-  text-align: center;
-  color: ${({ theme }) => theme.colors.semantic.text.sub};
-  padding: 2rem 0;
-  font-size: 1.1rem;
-`;
-
-const Section = styled.section`
-  padding: ${({ theme }) => theme.spacing.spacing4} ${({ theme }) => theme.spacing.spacing3};
-  width: 100%;
-  background-color: ${({ theme }) => theme.colors.semantic.background.default};
-`;
-
-const Title = styled.h3`
-  font-size: ${({ theme }) => theme.typography.title2Bold.fontSize};
-  font-weight: ${({ theme }) => theme.typography.title2Bold.fontWeight};
-  line-height: ${({ theme }) => theme.typography.title2Bold.lineHeight};
-  color: ${({ theme }) => theme.colors.semantic.text.default};
-  margin: 0px;
-  width: 100%;
-  text-align: left;
-`;
-
-const CatContainer = styled.div`
-  border-radius: ${({ theme }) => theme.spacing.spacing2};
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
-
-const GenerationGroup = styled.div`
-  display: flex;
-  gap: ${({ theme }) => theme.spacing.spacing2};
-  margin-bottom: ${({ theme }) => theme.spacing.spacing6};
-`;
-
-const FilterGroup = styled.div`
-  display: flex;
-  gap: ${({ theme }) => theme.spacing.spacing2};
-  margin-bottom: ${({ theme }) => theme.spacing.spacing6};
-`;
-
-const Button = styled.button<ButtonProps>`
-  padding: ${({ theme }) => theme.spacing.spacing3};
-  background-color: ${({ theme, isActive }) =>
-    isActive ? theme.colors.colorScale.blue[700] : theme.colors.colorScale.gray[100]};
-  border: 1px solid ${({ theme }) => theme.colors.colorScale.gray[300]};
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const RankContainer = styled.div`
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: ${({ theme }) => theme.spacing.spacing7} ${({ theme }) => theme.spacing.spacing2};
-`;
-
-const RankItem = styled.div`
-  width: 100%;
-  position: relative;
-  cursor: pointer;
-`;
-
-const RankNumber = styled.span`
-  position: absolute;
-  z-index: 2;
-  width: 1.25rem;
-  height: 1.25rem;
-  border-radius: 4px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: ${({ theme }) => theme.typography.subtitle2Bold.fontSize};
-  line-height: 1;
-  font-weight: ${({ theme }) => theme.typography.subtitle2Bold.fontWeight};
-  top: ${({ theme }) => theme.spacing.spacing1};
-  left: ${({ theme }) => theme.spacing.spacing1};
-  color: rgb(255, 255, 255);
-  background-color: ${({ theme }) => theme.colors.colorScale.red[600]};
-`;
-
-const Image = styled.img`
-  width: 100%;
-  object-fit: cover;
-  object-position: center center;
-  border-radius: 4px;
-  overflow: hidden;
-`;
-
-const ItemContainer = styled.div`
-  margin: 0;
-  padding: 0;
-  font-size: 100%;
-  font: inherit;
-  vertical-align: baseline;
-  box-sizing: border-box;
-`;
-
-const ItemBrandName = styled.p`
-  font-size: ${({ theme }) => theme.typography.body2Regular.fontSize};
-  font-weight: ${({ theme }) => theme.typography.body2Regular.fontWeight};
-  line-height: ${({ theme }) => theme.typography.body2Regular.lineHeight};
-  color: ${({ theme }) => theme.colors.semantic.text.sub};
-  margin: 0px;
-  text-align: left;
-`;
-
-const ItemName = styled.h6`
-  font-size: ${({ theme }) => theme.typography.body2Regular.fontSize};
-  font-weight: ${({ theme }) => theme.typography.body2Regular.fontWeight};
-  line-height: ${({ theme }) => theme.typography.body2Regular.lineHeight};
-  color: ${({ theme }) => theme.colors.semantic.text.default};
-  margin: 0px;
-  text-align: left;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-`;
-
-const ItemPrice = styled.p`
-  font-size: ${({ theme }) => theme.typography.body1Bold.fontSize};
-  font-weight: ${({ theme }) => theme.typography.body1Bold.fontWeight};
-  line-height: ${({ theme }) => theme.typography.body1Bold.lineHeight};
-  color: ${({ theme }) => theme.colors.semantic.text.default};
-  margin: 0px;
-  text-align: left;
-  word-break: break-word;
-`;
-
-const ToggleButton = styled.button`
-  width: 100%;
-  padding: ${({ theme }) => theme.spacing.spacing5};
-  border-radius: 4px;
-  border: 1px solid ${({ theme }) => theme.colors.colorScale.gray[300]};
-`;
