+++ src/page/Home/hooks/useThemes.ts
@@ -1,13 +1,8 @@
 import { requests } from '@/api/requests';
+import type { ThemeInfo } from '@/types';
 import { useEffect, useState } from 'react';
 
-export interface ThemeInfo {
-  themeId: number;
-  name: string;
-  image: string;
-}
-
-const useTheme = () => {
+const useThemes = () => {
   const [themes, setThemes] = useState<ThemeInfo[]>([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(false);
@@ -30,4 +25,4 @@ const useTheme = () => {
   return { themes, loading, error };
 };
 
-export default useTheme;
+export default useThemes;
