+++ src/page/Order/components/ReceiverInfoArray.tsx
@@ -1,5 +1,5 @@
+import type { OrderInfoValues } from '@/types';
 import styled from '@emotion/styled';
-import type { OrderInfoValues } from '..';
 
 interface ReceiverInfoArrayProps {
   receiverInfos: OrderInfoValues['receiverInfos'];
