+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './user';
+export * from './product';
+export * from './order';
+export * from './theme';
