+++ src/types/order.ts
@@ -1,0 +1,5 @@
+export interface OrderInfoValues {
+  message: string;
+  name: string;
+  receiverInfos: { name: string; phoneNumber: string; quantity: number }[];
+}
