+++ src/types/user.ts
@@ -1,0 +1,10 @@
+export interface UserInfoProps {
+  username: { value: string };
+  password: { value: string };
+}
+
+export interface UserInfoData {
+  name: string;
+  email: string;
+  authToken: string;
+}
